[compiler][ez] Move compiler gating tests
Move all gating tests to `gating/`
