--- compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/gating-test.js
@@ -1,14 +1,0 @@
-// @gating @compilationMode(annotation)
-function Bar(props) {
-  'use forget';
-  return <div>{props.bar}</div>;
-}
-
-function NoForget(props) {
-  return <Bar>{props.noForget}</Bar>;
-}
-
-function Foo(props) {
-  'use forget';
-  return <Foo>{props.bar}</Foo>;
-}
