+++ compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/gating/codegen-instrument-forget-gating-test.expect.md
@@ -15,8 +15,21 @@ function NoForget(props) {
 
 function Foo(props) {
   'use forget';
-  return <Foo>{props.bar}</Foo>;
+  if (props.bar < 0) {
+    return props.children;
+  }
+  return (
+    <Foo bar={props.bar - 1}>
+      <NoForget />
+    </Foo>
+  );
 }
+
+global.DEV = true;
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval('Foo'),
+  params: [{bar: 2}],
+};
 
 ```
 
@@ -29,7 +42,7 @@ import { c as _c } from "react/compiler-runtime"; // @enableEmitInstrumentForget
 const Bar = isForgetEnabled_Fixtures()
   ? function Bar(props) {
       "use forget";
-      if (__DEV__ && shouldInstrument)
+      if (DEV && shouldInstrument)
         useRenderCounter("Bar", "/codegen-instrument-forget-gating-test.ts");
       const $ = _c(2);
       let t0;
@@ -53,23 +66,50 @@ function NoForget(props) {
 const Foo = isForgetEnabled_Fixtures()
   ? function Foo(props) {
       "use forget";
-      if (__DEV__ && shouldInstrument)
+      if (DEV && shouldInstrument)
         useRenderCounter("Foo", "/codegen-instrument-forget-gating-test.ts");
-      const $ = _c(2);
-      let t0;
-      if ($[0] !== props.bar) {
-        t0 = <Foo>{props.bar}</Foo>;
-        $[0] = props.bar;
-        $[1] = t0;
-      } else {
-        t0 = $[1];
+      const $ = _c(3);
+      if (props.bar < 0) {
+        return props.children;
       }
-      return t0;
+
+      const t0 = props.bar - 1;
+      let t1;
+      if ($[0] === Symbol.for("react.memo_cache_sentinel")) {
+        t1 = <NoForget />;
+        $[0] = t1;
+      } else {
+        t1 = $[0];
+      }
+      let t2;
+      if ($[1] !== t0) {
+        t2 = <Foo bar={t0}>{t1}</Foo>;
+        $[1] = t0;
+        $[2] = t2;
+      } else {
+        t2 = $[2];
+      }
+      return t2;
     }
   : function Foo(props) {
       "use forget";
-      return <Foo>{props.bar}</Foo>;
+      if (props.bar < 0) {
+        return props.children;
+      }
+      return (
+        <Foo bar={props.bar - 1}>
+          <NoForget />
+        </Foo>
+      );
     };
+
+global.DEV = true;
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval("Foo"),
+  params: [{ bar: 2 }],
+};
 
 ```
       
+### Eval output
+(kind: ok) <div></div>
