+++ compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/gating/gating-test-export-default-function.js
@@ -1,14 +1,19 @@
 // @gating @compilationMode(annotation)
-export function Bar(props) {
+export default function Bar(props) {
   'use forget';
   return <div>{props.bar}</div>;
 }
 
-export function NoForget(props) {
+function NoForget(props) {
   return <Bar>{props.noForget}</Bar>;
 }
 
-export function Foo(props) {
+function Foo(props) {
   'use forget';
   return <Foo>{props.bar}</Foo>;
 }
+
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval('Bar'),
+  params: [{bar: 2}],
+};
