+++ compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/gating/gating-test-export-function-and-default.expect.md
@@ -14,8 +14,20 @@ function NoForget(props) {
 
 function Foo(props) {
   'use forget';
-  return <Foo>{props.bar}</Foo>;
+  if (props.bar < 0) {
+    return props.children;
+  }
+  return (
+    <Foo bar={props.bar - 1}>
+      <NoForget />
+    </Foo>
+  );
 }
+
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval('Bar'),
+  params: [{bar: 2}],
+};
 
 ```
 
@@ -50,21 +62,47 @@ function NoForget(props) {
 const Foo = isForgetEnabled_Fixtures()
   ? function Foo(props) {
       "use forget";
-      const $ = _c(2);
-      let t0;
-      if ($[0] !== props.bar) {
-        t0 = <Foo>{props.bar}</Foo>;
-        $[0] = props.bar;
-        $[1] = t0;
-      } else {
-        t0 = $[1];
+      const $ = _c(3);
+      if (props.bar < 0) {
+        return props.children;
       }
-      return t0;
+
+      const t0 = props.bar - 1;
+      let t1;
+      if ($[0] === Symbol.for("react.memo_cache_sentinel")) {
+        t1 = <NoForget />;
+        $[0] = t1;
+      } else {
+        t1 = $[0];
+      }
+      let t2;
+      if ($[1] !== t0) {
+        t2 = <Foo bar={t0}>{t1}</Foo>;
+        $[1] = t0;
+        $[2] = t2;
+      } else {
+        t2 = $[2];
+      }
+      return t2;
     }
   : function Foo(props) {
       "use forget";
-      return <Foo>{props.bar}</Foo>;
+      if (props.bar < 0) {
+        return props.children;
+      }
+      return (
+        <Foo bar={props.bar - 1}>
+          <NoForget />
+        </Foo>
+      );
     };
+
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval("Bar"),
+  params: [{ bar: 2 }],
+};
 
 ```
       
+### Eval output
+(kind: ok) <div>2</div>
