+++ compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/gating/gating-test-export-function.js
@@ -1,10 +1,10 @@
 // @gating @compilationMode(annotation)
-export default function Bar(props) {
+export function Bar(props) {
   'use forget';
   return <div>{props.bar}</div>;
 }
 
-function NoForget(props) {
+export function NoForget(props) {
   return <Bar>{props.noForget}</Bar>;
 }
 
@@ -12,3 +12,8 @@ export function Foo(props) {
   'use forget';
   return <Foo>{props.bar}</Foo>;
 }
+
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval('Bar'),
+  params: [{bar: 2}],
+};
