+++ compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/gating/gating-test.expect.md
@@ -3,7 +3,7 @@
 
 ```javascript
 // @gating @compilationMode(annotation)
-export default function Bar(props) {
+function Bar(props) {
   'use forget';
   return <div>{props.bar}</div>;
 }
@@ -12,10 +12,15 @@ function NoForget(props) {
   return <Bar>{props.noForget}</Bar>;
 }
 
-export function Foo(props) {
+function Foo(props) {
   'use forget';
   return <Foo>{props.bar}</Foo>;
 }
+
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval('Bar'),
+  params: [{bar: 2}],
+};
 
 ```
 
@@ -42,13 +47,11 @@ const Bar = isForgetEnabled_Fixtures()
       "use forget";
       return <div>{props.bar}</div>;
     };
-export default Bar;
 
 function NoForget(props) {
   return <Bar>{props.noForget}</Bar>;
 }
-
-export const Foo = isForgetEnabled_Fixtures()
+const Foo = isForgetEnabled_Fixtures()
   ? function Foo(props) {
       "use forget";
       const $ = _c(2);
@@ -67,5 +70,12 @@ export const Foo = isForgetEnabled_Fixtures()
       return <Foo>{props.bar}</Foo>;
     };
 
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval("Bar"),
+  params: [{ bar: 2 }],
+};
+
 ```
       
+### Eval output
+(kind: ok) <div>2</div>
