+++ compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/gating/multi-arrow-expr-export-gating-test.expect.md
@@ -3,14 +3,21 @@
 
 ```javascript
 // @gating
-const ErrorView = (error, _retry) => <MessageBox error={error}></MessageBox>;
+import {Stringify} from 'shared-runtime';
 
-export default Renderer = props => (
-  <Foo>
-    <Bar></Bar>
+const ErrorView = (error, _retry) => <Stringify error={error}></Stringify>;
+
+export const Renderer = props => (
+  <div>
+    <span></span>
     <ErrorView></ErrorView>
-  </Foo>
+  </div>
 );
+
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval('Renderer'),
+  params: [{}],
+};
 
 ```
 
@@ -19,12 +26,14 @@ export default Renderer = props => (
 ```javascript
 import { isForgetEnabled_Fixtures } from "ReactForgetFeatureFlag";
 import { c as _c } from "react/compiler-runtime"; // @gating
+import { Stringify } from "shared-runtime";
+
 const ErrorView = isForgetEnabled_Fixtures()
   ? (error, _retry) => {
       const $ = _c(2);
       let t0;
       if ($[0] !== error) {
-        t0 = <MessageBox error={error} />;
+        t0 = <Stringify error={error} />;
         $[0] = error;
         $[1] = t0;
       } else {
@@ -32,18 +41,18 @@ const ErrorView = isForgetEnabled_Fixtures()
       }
       return t0;
     }
-  : (error, _retry) => <MessageBox error={error}></MessageBox>;
+  : (error, _retry) => <Stringify error={error}></Stringify>;
 
-export default Renderer = isForgetEnabled_Fixtures()
+export const Renderer = isForgetEnabled_Fixtures()
   ? (props) => {
       const $ = _c(1);
       let t0;
       if ($[0] === Symbol.for("react.memo_cache_sentinel")) {
         t0 = (
-          <Foo>
-            <Bar />
+          <div>
+            <span />
             <ErrorView />
-          </Foo>
+          </div>
         );
         $[0] = t0;
       } else {
@@ -52,11 +61,17 @@ export default Renderer = isForgetEnabled_Fixtures()
       return t0;
     }
   : (props) => (
-      <Foo>
-        <Bar></Bar>
+      <div>
+        <span></span>
         <ErrorView></ErrorView>
-      </Foo>
+      </div>
     );
+export const FIXTURE_ENTRYPOINT = {
+  fn: eval("Renderer"),
+  params: [{}],
+};
 
 ```
       
+### Eval output
+(kind: ok) <div><span></span><div>{"error":{}}</div></div>
