--- compiler/packages/babel-plugin-react-compiler/src/__tests__/fixtures/compiler/multi-arrow-expr-export-gating-test.js
@@ -1,9 +1,0 @@
-// @gating
-const ErrorView = (error, _retry) => <MessageBox error={error}></MessageBox>;
-
-export const Renderer = props => (
-  <Foo>
-    <Bar></Bar>
-    <ErrorView></ErrorView>
-  </Foo>
-);
